requirements.txt
pypdf
openai
faiss-cpu
numpy